PlayerResignations => PlayerSet
I think that PlayerSet is actually a more useful abstraction than a
specific PlayerResignation code
