+++ lttcore/src/lib.rs
@@ -11,8 +11,8 @@ extern crate derive_builder;
 pub mod player;
 pub use player::Player;
 
-pub mod player_resignations;
-pub use player_resignations::PlayerResignations;
+pub mod player_set;
+pub use player_set::PlayerSet;
 
 pub mod number_of_players;
 pub use number_of_players::NumberOfPlayers;
